requirements.txt
Flask==2.3.2
Pillow==10.0.0
opencv-python-headless==4.8.0.74
numpy==1.24.3
pytesseract==0.3.10
transformers==4.31.0
torch==2.0.1